configured feat-auth on auth service mvc and resource

+++ src/main/java/com/example/demo/db/reposiroty/StoreRepository.java
@@ -7,5 +7,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
  * @author Sattya
  * create at 1/28/2024 5:27 PM
  */
-public interface StoreRepository extends JpaRepository<StoreEntity,Integer> {
+public interface StoreRepository extends JpaRepository<StoreEntity,Long> {
 }
