+++ src/main/java/com/example/demo/db/reposiroty/SupplierRepository.java
@@ -7,5 +7,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
  * @author Sattya
  * create at 1/28/2024 5:26 PM
  */
-public interface SupplierRepository extends JpaRepository<SupplierEntity,Integer> {
+public interface SupplierRepository extends JpaRepository<SupplierEntity,Long> {
 }
